requirements.txt
websockets
janus
Pillow
tesserocr
scikit-image
numpy
PyQt5
